Split EconomyModule into an event handler (still called EconomyModule) and a database handler (named EconomyInstance, located within shared/economydata.py)
